added better testing

+++ inc/DiffieHellman.h
@@ -12,8 +12,8 @@ class DiffieHellman {
 public:
     static void getDHB(uint8_t *preSharedKey,uint8_t *publicKey,uint8_t *privateKey);
     static void getSessionKey(uint8_t *preSharedKey,uint8_t *dha, uint8_t *privateKey,uint8_t *sessionKey);
-    static uint64_t powerMod(uint64_t a,uint64_t b,uint64_t prime);
-    static uint64_t multiplyMod(uint64_t a, uint64_t b, uint64_t prime);
+    static uint32_t powerMod(uint32_t a,uint32_t b,uint32_t prime);
+    static uint32_t multiplyMod(uint32_t a, uint32_t b, uint32_t prime);
 };
